Settle the two-level cell numbering at build finish instead of on the first lookup
`offsetInternal` no longer checks whether the prefix sums exist. The caller
computes them with `computeBucketPrefix` once the buckets stop growing, and
`HashJoin::onBuildPhaseFinish` does that for its two-level maps, after
`ConcurrentHashJoin` has merged the per-slot buckets into one map. That removes
an acquire load from every offset read on the probe path, and the table drops
the `std::once_flag` and the inline prefix array in favour of a vector that
exists only once it is computed.

`bits_for_bucket` is an `Int32`, and `bucketFor` stays public: it names the
sub-table of a key, which for fixed-range storage is not the bucket the key
routes to.

+++ src/Common/HashTable/PartitionedFixedHashMap.h
@@ -16,7 +16,7 @@ struct IsFixedRangeTable<FixedHashMap<Key, Mapped, Cell, FixedHashTableStoredSiz
 /// under, so the cells, their offsets and iteration are those of the plain `FixedHashMap`. Keys are
 /// routed by the cache line of their cell, not by the key's high bits, so a dense key range spreads
 /// over the buckets. With `bits_for_bucket = 0` it is the plain map with no routing at all.
-template <typename Key, typename Mapped, size_t size_bits = sizeof(Key) * 8, size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
+template <typename Key, typename Mapped, size_t size_bits = sizeof(Key) * 8, Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
 using PartitionedFixedHashMap = TwoLevelHashTable<
     Key,
     FixedHashMapCell<Key, Mapped>,
