+++ src/Common/HashTable/PartitionedFixedHashSet.h
@@ -12,7 +12,7 @@ struct IsFixedRangeTable<FixedHashSet<Key, Allocator, size_bits>> : std::true_ty
 
 /// Set counterpart of `PartitionedFixedHashMap`, for a caller that only records which keys exist.
 /// The cell holds nothing but the presence flag, so routing uses the cache line of that narrower cell.
-template <typename Key, size_t size_bits = sizeof(Key) * 8, size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
+template <typename Key, size_t size_bits = sizeof(Key) * 8, Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET>
 using PartitionedFixedHashSet = TwoLevelHashTable<
     Key,
     FixedHashTableCell<Key>,
