+++ src/Common/HashTable/TwoLevelHashMap.h
@@ -12,7 +12,7 @@ template
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
     template <typename ...> typename ImplTable = HashMapTable,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
 >
 class TwoLevelHashMapTable
     : public TwoLevelHashTable<Key, Cell, Hash, Grower, Allocator, ImplTable<Key, Cell, Hash, Grower, Allocator>, bits_for_bucket>
@@ -47,7 +47,7 @@ template
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
     template <typename ...> typename ImplTable = HashMapTable,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
 >
 using TwoLevelHashMap = TwoLevelHashMapTable<Key, HashMapCell<Key, Mapped, Hash>, Hash, Grower, Allocator, ImplTable, bits_for_bucket>;
 
@@ -60,7 +60,7 @@ template
     typename Grower = TwoLevelHashTableGrower<>,
     typename Allocator = HashTableAllocator,
     template <typename ...> typename ImplTable = HashMapTable,
-    size_t bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
+    Int32 bits_for_bucket = DEFAULT_BITS_FOR_BUCKET
 >
 using TwoLevelHashMapWithSavedHash
     = TwoLevelHashMapTable<Key, HashMapCellWithSavedHash<Key, Mapped, Hash>, Hash, Grower, Allocator, ImplTable, bits_for_bucket>;
